blog change to mongodb

+++ business/src/main/java/hit/go/forward/business/database/mongo/MongoDB.java
@@ -17,6 +17,7 @@ import org.bson.types.ObjectId;
 
 import hit.go.forward.common.entity.blog.Blog;
 import hit.go.forward.common.entity.mongo.BlogDocument;
+import hit.go.forward.common.entity.user.User;
 import hit.go.forward.common.util.blog.BlogEntityUtil;
 
 public class MongoDB {
@@ -44,11 +45,29 @@ public class MongoDB {
         collection.insertOne(BlogEntityUtil.toDocument(blog));
     }
 
-    public static void deleteById(String id) {
-        collection.deleteOne(Filters.eq("_id", new ObjectId(id)));
+    public static boolean deleteById(String id) {
+        return collection.deleteOne(Filters.eq("_id", new ObjectId(id))).wasAcknowledged();
     }
 
-    public static Blog getBlogById(String id, String userId, String level) {
+    public static boolean deleteById(String id, String userId) {
+        return collection.deleteOne(Filters.and(Filters.eq("_id", new ObjectId(id)), Filters.eq("userId", userId))).wasAcknowledged();
+    }
+
+    public static boolean updateBlog(Blog blog) {
+        return collection.updateOne(Filters.eq("_id", new ObjectId(blog.getId())), BlogEntityUtil.toDocument(blog)).wasAcknowledged();
+    }
+
+    public static boolean updateBlog(Blog blog, String userId) {
+        return collection.updateOne(Filters.and(Filters.eq("_id", new ObjectId(blog.getId())), Filters.eq("userId", userId)), BlogEntityUtil.toDocument(blog)).wasAcknowledged();
+    }
+
+    public static boolean increaseField(String blogId, String field) {
+        Document doc = collection.find(Filters.eq("_id", new ObjectId(blogId))).first();
+        doc.put(field, ((Integer)doc.get(field)) + 1);
+        return collection.updateOne(Filters.eq("_id", new ObjectId(blogId)), doc).wasAcknowledged();
+    }
+
+    public static Blog getBlogById(String id, String userId, Integer level) {
         FindIterable<Document> result = filterByUser(collection.find(Filters.eq("_id", new ObjectId(id))), userId, level);
         Document doc = result.first();
         Blog blog = null;
@@ -60,11 +79,20 @@ public class MongoDB {
         return blog;
     }
 
-    public static List<Document> getBlogByType(Integer type, String level, String userId, Integer start, Integer length) {
+    public static List<Document> getBlogByType(Integer type, Integer level, String userId, Integer start, Integer length) {
         FindIterable<Document> result = filterByUserPage(collection.find(Filters.eq("type", type)), userId, level, start, length);
         if (result == null) return null;
 
-        Iterator<Document> itr = result.iterator();
+        return docItrToList(result.iterator());
+    }
+
+    public static List<Document> queryBlogByUser(String userId, Integer start, Integer length) {
+        FindIterable<Document> result = collection.find(Filters.eq("userId", userId)).sort(Filters.eq("_id", 1)).skip(start).limit(length);
+        if (result == null) return null;
+        return docItrToList(result.iterator());
+    }
+
+    private static List<Document> docItrToList(Iterator<Document> itr) {
         List<Document> list = new ArrayList<>();
         while (itr.hasNext()) {
             list.add(itr.next());
@@ -72,14 +100,15 @@ public class MongoDB {
         return list;
     }
 
-    private static FindIterable<Document> filterByUser(FindIterable<Document> docs, String userId, String type) {
-        switch (type) {
-            case "admin":
+    private static FindIterable<Document> filterByUser(FindIterable<Document> docs, String userId, Integer userType) {
+        switch (userType) {
+            case User.TYPE_ADMIN:
                 break;
-            case "user":
+            case User.TYPE_STUDENT:
+            case User.TYPE_TEACHER:
                 docs = docs.filter(Filters.or(Filters.eq("userId", userId), Filters.eq("status", Blog.STATUS_RELEASED)));
                 break;
-            case "visitor":
+            case User.TYPE_VISITOR:
                 docs = docs.filter(Filters.eq("status", Blog.STATUS_RELEASED));
                 break;
             default:
@@ -89,9 +118,9 @@ public class MongoDB {
         return docs;
     }
 
-    private static FindIterable<Document> filterByUserPage(FindIterable<Document> docs, String userId, String type, Integer start, Integer length) {
+    private static FindIterable<Document> filterByUserPage(FindIterable<Document> docs, String userId, Integer userType, Integer start, Integer length) {
         if (start == null || length == null) return null;
-        return filterByUser(docs, userId, type).sort(Filters.eq("id", 1)).skip(start).limit(length);
+        return filterByUser(docs, userId, userType).sort(Filters.eq("_id", 1)).skip(start).limit(length);
     }
